split out modules per major type and add missing docs
