+++ src/lib.rs
@@ -19,11 +19,20 @@ mod elem_type;
 pub use elem_type::{ElemType, StackType};
 mod an_elem;
 pub use an_elem::{AnElem, AnElemError};
+mod an_elem_return;
+pub use an_elem_return::Return;
 mod location;
 pub use location::{ArgumentIndex, LineNo};
 mod stack;
 pub use stack::{Stack, StackError};
+
 mod types;
+pub use types::empty::Empty;
+pub use types::context::{Context, ContextError};
+pub use types::type_id::TypeId;
+pub use types::type_id::map::{TypeIdMap, TypeIdMapError};
+pub use types::{Type, TypeError};
+
 mod elems;
 pub use elems::{Elems, ElemsPopError};
 mod elems_singleton;
@@ -32,8 +41,26 @@ mod elems_or;
 pub use elems_or::Or;
 mod elems_all;
 pub use elems_all::AllElems;
-mod types_scratch;
-pub use types_scratch::{IList, IOList};
+mod elems_input;
+pub use elems_input::IElems;
+mod elems_input_output;
+pub use elems_input_output::IOElems;
+mod elems_input_output_singleton;
+pub use elems_input_output_singleton::ReturnSingleton;
+mod elems_input_output_or;
+pub use elems_input_output_or::ReturnOr;
+mod elems_list;
+pub use elems_list::IsList;
+mod elems_list_nil;
+pub use elems_list_nil::Nil;
+mod elems_list_cons;
+pub use elems_list_cons::{Cons, IterCons};
+mod elems_list_input;
+pub use elems_list_input::IList;
+mod elems_list_input_output;
+pub use elems_list_input_output::IOList;
+mod elems_list_input_output_cons;
+pub use elems_list_input_output_cons::ConsOut;
 mod json_template;
 pub use json_template::{TMap, TValue, TValueRunError, Template};
 mod query;
